13/09/2023/5:34PM

+++ src/main/java/com/driver/service/impl/OrderServiceImpl.java
@@ -1,5 +1,6 @@
 package com.driver.service.impl;
 
+import com.driver.Transformers.OrderTransformer;
 import com.driver.io.entity.OrderEntity;
 import com.driver.io.repository.OrderRepository;
 import com.driver.service.OrderService;
@@ -37,30 +38,25 @@ public class OrderServiceImpl implements OrderService
     {
         OrderEntity orderEntity=orderRepository.findByOrderId(orderId);
 
-        OrderDto orderDto=new OrderDto();
-
-        orderDto.setId(orderEntity.getId());
-        orderDto.setUserId(orderEntity.getUserId());
-        orderDto.setStatus(orderEntity.isStatus());
-        orderDto.setItems(orderEntity.getItems());
-        orderDto.setCost(orderEntity.getCost());
-        orderDto.setOrderId(orderEntity.getOrderId());
+        OrderDto orderDto=OrderTransformer.orderDtoFromOrderEntity(orderEntity);
         return orderDto;
     }
 
     @Override
     public OrderDto updateOrderDetails(String orderId, OrderDto order) throws Exception
     {
-        OrderEntity orderEntity=new OrderEntity();
-        orderEntity.setUserId(order.getUserId());
-        orderEntity.setItems(order.getItems());
-        orderEntity.setCost(order.getCost());
 
-        orderEntity=orderRepository.save(orderEntity);
+        OrderEntity orderEntity= orderRepository.findByOrderId(orderId);
+        long id=orderEntity.getId();
+        if(orderEntity==null)throw  new Exception();
+     orderEntity =OrderTransformer.orderEntityFromOrderDto(order);
 
-        order.setId(orderEntity.getId());
-        order.setOrderId(orderEntity.getOrderId());
-        order.setStatus(orderEntity.isStatus());
+     orderEntity.setId(id);
+
+    orderEntity = orderRepository.save(orderEntity);
+
+    order =OrderTransformer.orderDtoFromOrderEntity(orderEntity);
+
         return order;
     }
 
@@ -68,6 +64,8 @@ public class OrderServiceImpl implements OrderService
     public void deleteOrder(String orderId) throws Exception
     {
         OrderEntity orderEntity= orderRepository.findByOrderId(orderId);
+        if(orderEntity==null)throw new Exception();
+
         long id=orderEntity.getId();
         orderRepository.deleteById(id);
         return;
@@ -79,13 +77,7 @@ public class OrderServiceImpl implements OrderService
         List<OrderDto>dtoList=new ArrayList<>();
         for(OrderEntity orderEntity: orderRepository.findAll())
         {
-            OrderDto orderDto=new OrderDto();
-            orderDto.setStatus(orderEntity.isStatus());
-            orderDto.setId(orderEntity.getId());
-            orderDto.setCost(orderEntity.getCost());
-            orderDto.setUserId(orderEntity.getUserId());
-            orderDto.setOrderId(orderEntity.getOrderId());
-            orderDto.setItems(orderEntity.getItems());
+            OrderDto orderDto= OrderTransformer.orderDtoFromOrderEntity(orderEntity);
         }
         return dtoList;
     }
